backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "content-index-backend"
version = "0.1.0"
description = "Backend API for content index page"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "python-multipart==0.0.19",
    "pydantic==2.10.4",
    "python-dotenv==1.0.1",
    "httpx==0.28.1",
    "azure-ai-ml==1.23.0",
    "azure-identity==1.19.0",
    "azure-ai-projects==1.0.0b10",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.28.0",
]
